+++ src/test/scala/com/johnsnowlabs/ml/logreg/I2b2DatasetLogRegTest.scala
@@ -1,14 +1,15 @@
 package com.johnsnowlabs.ml.logreg
 
 
+import com.johnsnowlabs.ml.common.EvaluationMetrics
 import com.johnsnowlabs.nlp.annotators.assertion.logreg.{SimpleTokenizer, Tokenizer, Windowing}
 import com.johnsnowlabs.nlp.embeddings.WordEmbeddings
 import org.apache.spark.ml.classification.LogisticRegression
 import org.apache.spark.ml.linalg.Vectors
 import org.apache.spark.sql.functions._
-import org.apache.spark.sql.{ColumnName, DataFrame, SparkSession}
+import org.apache.spark.sql.{DataFrame, SparkSession}
 
-object I2b2DatasetLogRegTest extends App with Windowing {
+object I2b2DatasetLogRegTest extends App with Windowing with EvaluationMetrics {
 
   override val before = 11
   override val after = 13
@@ -41,29 +42,17 @@ object I2b2DatasetLogRegTest extends App with Windowing {
   println("testDsSize: " +  testDataset.count)
 
   val model = train(trainDataset.cache())
-  case class TpFnFp(tp: Int, fn: Int, fp: Int)
 
   // Compute raw scores on the test set.
   val result = model.transform(testDataset.cache())
 
-  val tpTnFp = result.map ({ r =>
-    if (r.getAs[Double]("prediction") == r.getAs[Double]("label")) TpFnFp(1, 0, 0)
-    else TpFnFp(0, 1, 1)
-  }).collect().reduce((t1, t2) => TpFnFp(t1.tp + t2.tp, t1.fn + t2.fn, t1.fp + t2.fp))
-
-  println(calcStat(tpTnFp.tp + tpTnFp.fn, tpTnFp.tp + tpTnFp.fp, tpTnFp.tp))
-
   val badGuys = result.filter(r => r.getAs[Double]("prediction") != r.getAs[Double]("label")).collect()
   println(badGuys)
 
-  val pred = result.select($"prediction").collect.map{ r =>
-    r.getAs[Double]("prediction")
-  }
+  val pred = result.select($"prediction").collect.map(_.getAs[Double]("prediction"))
+  val gold = result.select($"label").collect.map(_.getAs[Double]("label"))
 
-  val gold = result.select($"label").collect.map{ r =>
-    r.getAs[Double]("label")
-  }
-
+  println(calcStat(pred, gold))
 
   println(confusionMatrix(pred, gold))
 
@@ -76,31 +65,10 @@ object I2b2DatasetLogRegTest extends App with Windowing {
     lr.fit(dataFrame)
   }
 
-  def calcStat(correct: Long, predicted: Long, predictedCorrect: Long): (Float, Float, Float) = {
-    val prec = predictedCorrect.toFloat / predicted
-    val rec = predictedCorrect.toFloat / correct
-    val f1 = 2 * prec * rec / (prec + rec)
-    (prec, rec, f1)
-  }
-
-  def confusionMatrix[T](predicted: Seq[T], gold: Seq[T]) = {
-    val labels = gold.distinct
-    import scala.collection.mutable.{Map => MutableMap}
-    val matrix : Map[T, MutableMap[T, Int]] =
-      labels.map(label => (label, MutableMap(labels.zip(Array.fill(labels.size)(0)): _*))).toMap
-
-    predicted.zip(gold).foreach { case (p, g) => matrix.get(p).get(g) += 1}
-
-    /* sanity check, the confusion matrix should contain as many elements as there were used during training / prediction */
-    assert(predicted.length ==matrix.map(map => map._2.values.sum).sum)
-    matrix
-  }
-
   // produces a org.apache.spark.ml.linalg.Vector
   def convertToVectorUdf = udf {(array: Array[Double]) =>
-      val tmp = Vectors.dense(array)
-      tmp
+      Vectors.dense(array)
   }
 
-  override val wordVectors: Option[WordEmbeddings] = reader.wordVectors
+  override lazy val wordVectors: Option[WordEmbeddings] = reader.wordVectors
 }
